added position element

+++ ozobot.k3dsa/xtend-gen/ozobot/k3dsa/OzobotAspect.java
@@ -29,38 +29,38 @@ public class OzobotAspect extends NamedElementAspect {
     _privk3_main(_self_, _self);;
   }
   
-  public static float xposition(final Ozobot _self) {
+  public static double xposition(final Ozobot _self) {
     final ozobot.k3dsa.OzobotAspectOzobotAspectProperties _self_ = ozobot.k3dsa.OzobotAspectOzobotAspectContext.getSelf(_self);
     Object result = null;
     result = _privk3_xposition(_self_, _self);;
-    return (float)result;
+    return (double)result;
   }
   
-  public static void xposition(final Ozobot _self, final float xposition) {
+  public static void xposition(final Ozobot _self, final double xposition) {
     final ozobot.k3dsa.OzobotAspectOzobotAspectProperties _self_ = ozobot.k3dsa.OzobotAspectOzobotAspectContext.getSelf(_self);
     _privk3_xposition(_self_, _self,xposition);;
   }
   
-  public static float yposition(final Ozobot _self) {
+  public static double yposition(final Ozobot _self) {
     final ozobot.k3dsa.OzobotAspectOzobotAspectProperties _self_ = ozobot.k3dsa.OzobotAspectOzobotAspectContext.getSelf(_self);
     Object result = null;
     result = _privk3_yposition(_self_, _self);;
-    return (float)result;
+    return (double)result;
   }
   
-  public static void yposition(final Ozobot _self, final float yposition) {
+  public static void yposition(final Ozobot _self, final double yposition) {
     final ozobot.k3dsa.OzobotAspectOzobotAspectProperties _self_ = ozobot.k3dsa.OzobotAspectOzobotAspectContext.getSelf(_self);
     _privk3_yposition(_self_, _self,yposition);;
   }
   
-  public static float orientation(final Ozobot _self) {
+  public static double orientation(final Ozobot _self) {
     final ozobot.k3dsa.OzobotAspectOzobotAspectProperties _self_ = ozobot.k3dsa.OzobotAspectOzobotAspectContext.getSelf(_self);
     Object result = null;
     result = _privk3_orientation(_self_, _self);;
-    return (float)result;
+    return (double)result;
   }
   
-  public static void orientation(final Ozobot _self, final float orientation) {
+  public static void orientation(final Ozobot _self, final double orientation) {
     final ozobot.k3dsa.OzobotAspectOzobotAspectProperties _self_ = ozobot.k3dsa.OzobotAspectOzobotAspectContext.getSelf(_self);
     _privk3_orientation(_self_, _self,orientation);;
   }
@@ -104,14 +104,14 @@ public class OzobotAspect extends NamedElementAspect {
     }
   }
   
-  protected static float _privk3_xposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
+  protected static double _privk3_xposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
     		if (m.getName().equals("getXposition") &&
     			m.getParameterTypes().length == 0) {
     				Object ret = m.invoke(_self);
     				if (ret != null) {
-    					return (float) ret;
+    					return (double) ret;
     				}		}
     	}
     } catch (Exception e) {
@@ -120,7 +120,7 @@ public class OzobotAspect extends NamedElementAspect {
     return _self_.xposition;
   }
   
-  protected static void _privk3_xposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final float xposition) {
+  protected static void _privk3_xposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final double xposition) {
     boolean setterCalled = false;
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
@@ -138,14 +138,14 @@ public class OzobotAspect extends NamedElementAspect {
     }
   }
   
-  protected static float _privk3_yposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
+  protected static double _privk3_yposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
     		if (m.getName().equals("getYposition") &&
     			m.getParameterTypes().length == 0) {
     				Object ret = m.invoke(_self);
     				if (ret != null) {
-    					return (float) ret;
+    					return (double) ret;
     				}		}
     	}
     } catch (Exception e) {
@@ -154,7 +154,7 @@ public class OzobotAspect extends NamedElementAspect {
     return _self_.yposition;
   }
   
-  protected static void _privk3_yposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final float yposition) {
+  protected static void _privk3_yposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final double yposition) {
     boolean setterCalled = false;
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
@@ -172,14 +172,14 @@ public class OzobotAspect extends NamedElementAspect {
     }
   }
   
-  protected static float _privk3_orientation(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
+  protected static double _privk3_orientation(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
     		if (m.getName().equals("getOrientation") &&
     			m.getParameterTypes().length == 0) {
     				Object ret = m.invoke(_self);
     				if (ret != null) {
-    					return (float) ret;
+    					return (double) ret;
     				}		}
     	}
     } catch (Exception e) {
@@ -188,7 +188,7 @@ public class OzobotAspect extends NamedElementAspect {
     return _self_.orientation;
   }
   
-  protected static void _privk3_orientation(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final float orientation) {
+  protected static void _privk3_orientation(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final double orientation) {
     boolean setterCalled = false;
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
