+++ ozobot.xdsml.ozobotl/src-gen/ozobot/xdsml/ozobotl/aspects/OzobotAspect.java
@@ -33,48 +33,48 @@ public class OzobotAspect extends NamedElementAspect {
 	;
 }
   
-  public static float xposition(final Ozobot _self) {
+  public static double xposition(final Ozobot _self) {
 	final ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectProperties _self_ = ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectContext
 			.getSelf(_self);
 	Object result = null;
 	result = _privk3_xposition(_self_, _self);
 	;
-	return (float) result;
+	return (double) result;
 }
   
-  public static void xposition(final Ozobot _self, final float xposition) {
+  public static void xposition(final Ozobot _self, final double xposition) {
 	final ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectProperties _self_ = ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectContext
 			.getSelf(_self);
 	_privk3_xposition(_self_, _self, xposition);
 	;
 }
   
-  public static float yposition(final Ozobot _self) {
+  public static double yposition(final Ozobot _self) {
 	final ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectProperties _self_ = ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectContext
 			.getSelf(_self);
 	Object result = null;
 	result = _privk3_yposition(_self_, _self);
 	;
-	return (float) result;
+	return (double) result;
 }
   
-  public static void yposition(final Ozobot _self, final float yposition) {
+  public static void yposition(final Ozobot _self, final double yposition) {
 	final ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectProperties _self_ = ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectContext
 			.getSelf(_self);
 	_privk3_yposition(_self_, _self, yposition);
 	;
 }
   
-  public static float orientation(final Ozobot _self) {
+  public static double orientation(final Ozobot _self) {
 	final ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectProperties _self_ = ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectContext
 			.getSelf(_self);
 	Object result = null;
 	result = _privk3_orientation(_self_, _self);
 	;
-	return (float) result;
+	return (double) result;
 }
   
-  public static void orientation(final Ozobot _self, final float orientation) {
+  public static void orientation(final Ozobot _self, final double orientation) {
 	final ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectProperties _self_ = ozobot.xdsml.ozobotl.aspects.OzobotAspectOzobotAspectContext
 			.getSelf(_self);
 	_privk3_orientation(_self_, _self, orientation);
@@ -124,14 +124,14 @@ public class OzobotAspect extends NamedElementAspect {
     }
   }
   
-  protected static float _privk3_xposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
+  protected static double _privk3_xposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
     		if (m.getName().equals("getXposition") &&
     			m.getParameterTypes().length == 0) {
     				Object ret = m.invoke(_self);
     				if (ret != null) {
-    					return (float) ret;
+    					return (double) ret;
     				}		}
     	}
     } catch (Exception e) {
@@ -140,7 +140,7 @@ public class OzobotAspect extends NamedElementAspect {
     return _self_.xposition;
   }
   
-  protected static void _privk3_xposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final float xposition) {
+  protected static void _privk3_xposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final double xposition) {
     boolean setterCalled = false;
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
@@ -158,14 +158,14 @@ public class OzobotAspect extends NamedElementAspect {
     }
   }
   
-  protected static float _privk3_yposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
+  protected static double _privk3_yposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
     		if (m.getName().equals("getYposition") &&
     			m.getParameterTypes().length == 0) {
     				Object ret = m.invoke(_self);
     				if (ret != null) {
-    					return (float) ret;
+    					return (double) ret;
     				}		}
     	}
     } catch (Exception e) {
@@ -174,7 +174,7 @@ public class OzobotAspect extends NamedElementAspect {
     return _self_.yposition;
   }
   
-  protected static void _privk3_yposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final float yposition) {
+  protected static void _privk3_yposition(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final double yposition) {
     boolean setterCalled = false;
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
@@ -192,14 +192,14 @@ public class OzobotAspect extends NamedElementAspect {
     }
   }
   
-  protected static float _privk3_orientation(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
+  protected static double _privk3_orientation(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self) {
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
     		if (m.getName().equals("getOrientation") &&
     			m.getParameterTypes().length == 0) {
     				Object ret = m.invoke(_self);
     				if (ret != null) {
-    					return (float) ret;
+    					return (double) ret;
     				}		}
     	}
     } catch (Exception e) {
@@ -208,7 +208,7 @@ public class OzobotAspect extends NamedElementAspect {
     return _self_.orientation;
   }
   
-  protected static void _privk3_orientation(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final float orientation) {
+  protected static void _privk3_orientation(final OzobotAspectOzobotAspectProperties _self_, final Ozobot _self, final double orientation) {
     boolean setterCalled = false;
     try {
     	for (java.lang.reflect.Method m : _self.getClass().getMethods()) {
